+++ fork/dynamic-cdn-webpack-plugin/src/index.js
@@ -223,7 +223,9 @@ class DynamicCdnWebpackPlugin {
 	addDependencies(contextPath, manifest, { env, requester }) {
 		for (const dependencyName of Object.keys(manifest)) {
 			const cdnConfig = manifest[dependencyName];
-			const cwd = resolvePkg(cdnConfig.name, { cwd: contextPath });
+			const cwd = resolvePkg(cdnConfig.name, {
+				version: cdnConfig.version,
+			});
 			if (!cwd) {
 				this.error(
 					'\n❌',
@@ -276,7 +278,9 @@ class DynamicCdnWebpackPlugin {
 
 	async addModule(contextPath, modulePath, { env, isOptional = false }) {
 		const isModuleExcluded =
-			this.exclude.includes(modulePath) || (this.only && !this.only.includes(modulePath));
+			this.exclude.includes(modulePath) ||
+			(this.only && !this.only.includes(modulePath)) ||
+			modulePath.startsWith('@types/');
 		if (isModuleExcluded) {
 			return false;
 		}
@@ -316,7 +320,7 @@ class DynamicCdnWebpackPlugin {
 				if (!this.directDependencies[modulePath]) {
 					this.directDependencies[modulePath] = this.modulesFromCdn[modulePath];
 				}
-				return this.modulesFromCdn[modulePath].var;
+				return this.modulesFromCdn[modulePath].var || true;
 			}
 
 			this.log(
@@ -401,7 +405,7 @@ class DynamicCdnWebpackPlugin {
 		this.modulesFromCdn[modulePath] = cdnConfig;
 		this.directDependencies[modulePath] = cdnConfig;
 		this.debug('\n✅', modulePath, version, `will be served by ${cdnConfig.url}`);
-		return cdnConfig.var;
+		return cdnConfig.var || true;
 	}
 
 	applyWebpackCore(compiler) {