+++ packages/datagrid/src/components/datagrid.component.stories.js
@@ -1,18 +1,15 @@
-/* eslint-disable no-console */
+/* eslint-disable no-console, react/prop-types */
 
 import React, { useState } from 'react';
 import { action } from '@storybook/addon-actions';
-import { IconsProvider } from '@talend/react-components';
 
 import DataGrid from '.';
 import DynamicDataGrid from '../../stories/DynamicDataGrid.component';
 import FasterDatagridComponent from '../../stories/FasterDatagrid.component';
-import ImmutableDataGrid from '../../stories/ImmutableDatagrid.component';
 import sample from '../../stories/sample.json';
 import sample2 from '../../stories/sample2.json';
 import sample3 from '../../stories/sample3.json';
 import sampleWithoutQuality from '../../stories/sampleWithoutQuality.json';
-import getComponent from '../../stories/getComponent';
 
 // eslint-disable-next-line no-irregular-whitespace
 sample.data[0].value.field0.value = `﻿﻿﻿﻿﻿﻿﻿  loreum lo
@@ -24,15 +21,18 @@ very loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
 sample.data[2].value.field0.value =
 	'very looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong value';
 
+sample.data[3].value.field0.value = 'multiple       spaces';
+
 export default {
 	title: 'Datagrid/Component',
-	decorators: [story => <div style={{ height: '90vh' }}>{story()}</div>],
+	decorators: [
+		story => <div style={{ height: '90vh', backgroundColor: 'lightGrey' }}>{story()}</div>,
+	],
 };
 
 export const Default = () => (
 	<DataGrid
 		data={sample}
-		getComponent={getComponent}
 		onFocusedCell={action('onFocusedCell')}
 		onFocusedColumn={action('onFocusedColumn')}
 		onVerticalScroll={event => console.log(event)}
@@ -40,12 +40,36 @@ export const Default = () => (
 		enableColResize={false}
 	/>
 );
+Default.parameters = {
+	chromatic: { disableSnapshot: false },
+};
 
-export const NoSubtype = () => (
+export const CustomRenderer = () => (
 	<DataGrid
-		columnsConf={{ hideSubType: true }}
 		data={sample}
-		getComponent={getComponent}
+		cellRenderer={props => <div>{`${props.value.value}`}&#128570;</div>}
+		headerRenderer={props => <div>{props.displayName} &#128126;</div>}
+		pinHeaderRenderer={() => <div>&#129302;</div>}
+	/>
+);
+
+export const WithSelection = () => (
+	<DataGrid
+		data={sample}
+		onFocusedCell={action('onFocusedCell')}
+		onFocusedColumn={action('onFocusedColumn')}
+		focusedColumnId="data.field2"
+	/>
+);
+WithSelection.parameters = {
+	chromatic: { disableSnapshot: false },
+};
+
+export const OnlyColumnName = () => (
+	<DataGrid
+		headerHeight={45}
+		columnsConf={{ hideSubType: true }}
+		data={sampleWithoutQuality}
 		onFocusedCell={action('onFocusedCell')}
 		onFocusedColumn={action('onFocusedColumn')}
 		onVerticalScroll={event => console.log(event)}
@@ -56,8 +80,8 @@ export const NoSubtype = () => (
 
 export const NoQuality = () => (
 	<DataGrid
+		headerHeight={55}
 		data={sampleWithoutQuality}
-		getComponent={getComponent}
 		onFocusedCell={action('onFocusedCell')}
 		onFocusedColumn={action('onFocusedColumn')}
 		onVerticalScroll={event => console.log(event)}
@@ -69,7 +93,6 @@ export const NoQuality = () => (
 export const ColumnsResizables = () => (
 	<DataGrid
 		data={sample}
-		getComponent={getComponent}
 		onFocusedCell={action('onFocusedCell')}
 		onFocusedColumn={action('onFocusedColumn')}
 		onVerticalScroll={event => console.log(event)}
@@ -80,7 +103,6 @@ export const ColumnsResizables = () => (
 export const StartIndexTo1 = () => (
 	<DataGrid
 		data={sample}
-		getComponent={getComponent}
 		startIndex={1}
 		onFocusedCell={action('onFocusedCell')}
 		onFocusedColumn={action('onFocusedColumn')}
@@ -92,7 +114,6 @@ export const StartIndexTo1 = () => (
 export const NoRowSpecificMessage = () => (
 	<DataGrid
 		data={[]}
-		getComponent={getComponent}
 		overlayNoRowsTemplate="Custom message"
 		onFocusedCell={action('onFocusedCell')}
 		onFocusedColumn={action('onFocusedColumn')}
@@ -125,15 +146,9 @@ export const DynamicChangeSchema = () => {
 				<div>
 					<input type="button" value="changestatus" onClick={this.changeState} />
 					Number of fields : {currentSample.schema.fields.length}
-					<IconsProvider
-						bundles={[
-							'https://statics-dev.cloud.talend.com/@talend/icons/6.1.4/dist/svg-bundle/all.svg',
-						]}
-					/>
 					<div style={{ height: '200px' }}>
 						<DataGrid
 							data={currentSample}
-							getComponent={getComponent}
 							onFocusedCell={action('onFocusedCell')}
 							onFocusedColumn={action('onFocusedColumn')}
 							onVerticalScroll={event => console.log(event)}
@@ -147,15 +162,9 @@ export const DynamicChangeSchema = () => {
 	return <WithLayout />;
 };
 
-export const DynamicChangeDataWithForceRedrawRows = () => <DynamicDataGrid forceRedraw />;
-DynamicChangeDataWithForceRedrawRows.storyName =
-	'@deprecated - dynamic change data with forceRedrawRows';
-
 export const DynamicChangeData = () => <DynamicDataGrid />;
 
 export const FasterDatagrid = () => <FasterDatagridComponent />;
-
-export const ImmutableData = () => <ImmutableDataGrid />;
 
 export const ControlledFocusedColumn = () => {
 	const [focusedColumnId, setFocusedColumnId] = useState('data.field2');
@@ -176,7 +185,6 @@ export const ControlledFocusedColumn = () => {
 			<input type="button" value={locked ? 'Unlock' : 'Lock'} onClick={() => setLocked(!locked)} />
 			<DataGrid
 				data={sample}
-				getComponent={getComponent}
 				focusedColumnId={focusedColumnId}
 				onFocusedCell={cell => {
 					if (!locked) {