+++ packages/playground/webpack.config.dev.js
@@ -17,6 +17,7 @@ const mockBackend = require('./mockBackend/server');
 function getPath(pkg) {
 	const pkgPath = resolve(pkg, { cwd: process.cwd() });
 	return pkgPath
+		.replace('main.js', '')
 		.replace('lib/index.js', '')
 		.replace('dist/bootstrap.js', '')
 		.replace('dist/TalendIcons.js', '');
@@ -27,8 +28,10 @@ function getVersion(pkg) {
 }
 
 const PKGS = [
+	'@talend/assets-api',
 	'@talend/design-tokens',
 	'@talend/design-system',
+	'@talend/design-tokens',
 	'@talend/react-bootstrap',
 	'@talend/react-components',
 	'@talend/react-containers',