+++ packages/dataviz/CHANGELOG.md
@@ -1,5 +1,37 @@
 # Changelog
 
+## 2.1.2
+
+### Patch Changes
+
+- 9222aa7fc: fix: use assets-api from CDN
+- Updated dependencies [9222aa7fc]
+  - @talend/assets-api@1.1.0
+
+## 2.1.1
+
+### Patch Changes
+
+- 20e907887: formating to comply with lint
+- Updated dependencies [82bb8c1c8]
+- Updated dependencies [e04e3910f]
+  - @talend/design-tokens@2.2.0
+  - @talend/react-components@6.46.2
+
+## 2.1.0
+
+### Minor Changes
+
+- d1815c0af: feat: Use @talend/assets-api to load topologies.
+
+## 2.0.0
+
+### Major Changes
+
+- 9f3cad9bd: feat(TAD-6145): add a LineChart component in dataviz
+
+  Migration: TooltipContent(Props) is renamed to KeyValueTooltip(Props)
+
 ## 1.0.6
 
 ### Patch Changes