+++ tools/scripts-config-jest/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @talend/scripts-config-jest
 
+## 10.1.0
+
+### Minor Changes
+
+- f038a450c: feat(jest): support react 17
+
+## 10.0.0
+
+### Major Changes
+
+- 2875eb2ce: Breaking change: design-system mocked component now generate a span
+
 ## 9.7.3
 
 ### Patch Changes